backend/pyproject.toml
[project]
name = "support-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "python-dotenv>=1.0",
  "supabase>=2.4.0",
]

[tool.uvicorn]
# Optional note: run via `uvicorn app.main:app --reload --port 8000`
